builder completado

+++ src/main/java/com/ucu/olimpiadas/Model/Weightlifting.java
@@ -3,10 +3,10 @@ package com.ucu.olimpiadas.Model;
 public class Weightlifting implements IDiscipline{
     private String name;
     private String category;
-    private int snatch_weight;
-    private int clean_and_jerk_weight;
+    private double snatch_weight;
+    private double clean_and_jerk_weight;
 
-    public Weightlifting(String name, String category, int snatch_weight, int clean_and_jerk_weight) {
+    public Weightlifting(String name, String category, double snatch_weight, double clean_and_jerk_weight) {
         this.name = name;
         this.category = category;
         this.snatch_weight = snatch_weight;
